[#95] refactor - 커스텀훅으로 제작, 안쓰는 코드 legacy로 이동
