+++ src/components/Legacy/CategoryList/CategoryList.module.scss
