+++ src/pages/MyCartPage/MyCartPage.js
@@ -1,5 +1,5 @@
 /* External dependencies */
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import classNames from 'classnames/bind';
 
 /* Internal dependencies */
@@ -9,37 +9,12 @@ import Header from 'components/MyDetailPage/Header';
 import CartList from 'components/MyDetailPage/CartPage/CartList';
 import CartFooter from 'components/MyDetailPage/CartPage/CartFooter';
 import { FaShoppingCart } from 'react-icons/fa';
-import { getProjectListInCartAPI } from 'api/cartAPI';
+import useCart from 'hooks/useCart';
 
 const cx = classNames.bind(styles);
 
 function MyCartPage() {
-  const [cartList, setCartList] = useState([]);
-  const [totalPrice, setTotalPrice] = useState(0);
-  let cartPrice = 0;
-
-  useEffect(() => {
-    handleCartList();
-  }, []);
-
-  const handleCartList = function () {
-    getProjectListInCartAPI().then(result => {
-      setCartList(result.data);
-      result.data.forEach(cart => {
-        cartPrice += cart.money;
-      });
-      setTotalPrice(cartPrice);
-    });
-  };
-
-  function onModify(cart_id) {
-    handleCartList();
-  }
-
-  function onRemove(cart_id) {
-    setCartList(cartList.filter(cart => cart.cart_id !== cart_id));
-  }
-
+  const { cartList, onModify, onRemove, totalPrice } = useCart();
   return (
     <div className={cx('mycart-page')}>
       <BackButton />
